feat : collide the ground

+++ server/background/entities/background_entity.go
@@ -8,8 +8,8 @@ type Ranges struct {
 }
 
 type Position struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 type Background struct {
